requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
aio-pika==9.5.4
httpx==0.28.1
cryptography==44.0.0
google-generativeai==0.8.4
python-dotenv==1.0.1
